--- apps/overstock/app/search/loading.tsx
@@ -1,15 +1,0 @@
-import Grid from 'components/grid';
-
-export default function Loading() {
-  return (
-    <Grid className="grid-cols-2 lg:grid-cols-3">
-      {Array(12)
-        .fill(0)
-        .map((_, index) => {
-          return (
-            <Grid.Item key={index} className="animate-pulse bg-neutral-100 dark:bg-neutral-800" />
-          );
-        })}
-    </Grid>
-  );
-}
