+++ apps/overstock/components/cart/actions.ts
@@ -1,118 +1,25 @@
-'use server';
+"use server";
 
-import { TAGS } from 'lib/constants';
-import { addToCart, createCart, getCart, removeFromCart, updateCart } from 'lib/shopify';
-import { revalidateTag } from 'next/cache';
-import { cookies } from 'next/headers';
-import { redirect } from 'next/navigation';
+import { reshapeCart, shopifyFetch } from "lib/shopify";
+import { getCartQuery } from "lib/shopify/queries/cart";
+import type { ShopifyCartOperation } from "lib/shopify/types";
+import { cookies } from "next/headers";
 
-export async function addItem(prevState: any, selectedVariantId: string | undefined) {
-  let cartId = cookies().get('cartId')?.value;
+export async function getCart() {
+  const value = cookies().get("cart")?.value;
 
-  if (!cartId || !selectedVariantId) {
-    return 'Error adding item to cart';
-  }
+  if (value) {
+    const cart = await shopifyFetch<ShopifyCartOperation>({
+      query: getCartQuery,
+      variables: { cartId: `gid://shopify/Cart/${value}` },
+    });
 
-  try {
-    await addToCart(cartId, [{ merchandiseId: selectedVariantId, quantity: 1 }]);
-    revalidateTag(TAGS.cart);
-  } catch (e) {
-    return 'Error adding item to cart';
-  }
-}
-
-export async function removeItem(prevState: any, merchandiseId: string) {
-  let cartId = cookies().get('cartId')?.value;
-
-  if (!cartId) {
-    return 'Missing cart ID';
-  }
-
-  try {
-    const cart = await getCart(cartId);
-
-    if (!cart) {
-      return 'Error fetching cart';
+    if (!cart.body.data.cart) {
+      return undefined;
     }
 
-    const lineItem = cart.lines.find((line) => line.merchandise.id === merchandiseId);
-
-    if (lineItem && lineItem.id) {
-      await removeFromCart(cartId, [lineItem.id]);
-      revalidateTag(TAGS.cart);
-    } else {
-      return 'Item not found in cart';
-    }
-  } catch (e) {
-    return 'Error removing item from cart';
-  }
-}
-
-export async function updateItemQuantity(
-  prevState: any,
-  payload: {
-    merchandiseId: string;
-    quantity: number;
-  }
-) {
-  let cartId = cookies().get('cartId')?.value;
-
-  if (!cartId) {
-    return 'Missing cart ID';
+    return reshapeCart(cart.body.data.cart);
   }
 
-  const { merchandiseId, quantity } = payload;
-
-  try {
-    const cart = await getCart(cartId);
-
-    if (!cart) {
-      return 'Error fetching cart';
-    }
-
-    const lineItem = cart.lines.find((line) => line.merchandise.id === merchandiseId);
-
-    if (lineItem && lineItem.id) {
-      if (quantity === 0) {
-        await removeFromCart(cartId, [lineItem.id]);
-      } else {
-        await updateCart(cartId, [
-          {
-            id: lineItem.id,
-            merchandiseId,
-            quantity
-          }
-        ]);
-      }
-    } else if (quantity > 0) {
-      // If the item doesn't exist in the cart and quantity > 0, add it
-      await addToCart(cartId, [{ merchandiseId, quantity }]);
-    }
-
-    revalidateTag(TAGS.cart);
-  } catch (e) {
-    console.error(e);
-    return 'Error updating item quantity';
-  }
-}
-
-export async function redirectToCheckout() {
-  let cartId = cookies().get('cartId')?.value;
-
-  if (!cartId) {
-    return 'Missing cart ID';
-  }
-
-  let cart = await getCart(cartId);
-
-  if (!cart) {
-    return 'Error fetching cart';
-  }
-
-  redirect(cart.checkoutUrl);
-}
-
-export async function createCartAndSetCookie() {
-  let cart = await createCart();
-  cookies().set('cartId', cart.id!);
+  return undefined;
 }
