--- apps/overstock/components/cart/close-cart.tsx
@@ -1,10 +1,0 @@
-import { XMarkIcon } from '@heroicons/react/24/outline';
-import clsx from 'clsx';
-
-export default function CloseCart({ className }: { className?: string }) {
-  return (
-    <div className="relative flex h-6 w-6 items-center justify-center rounded-md border border-neutral-200 text-black transition-colors dark:border-neutral-700 dark:text-white">
-      <XMarkIcon className={clsx('h-6 transition-all ease-in-out hover:scale-110', className)} />
-    </div>
-  );
-}
